refactor(breadcrumbs): address first batch of overflow PR review
- Inline `BreadcrumbsOverlayMixin` logic into `vaadin-breadcrumbs-overlay.js`
  and delete the standalone mixin files
- Drop the empty `vaadin-breadcrumbs-overlay-base-styles.js` / `.d.ts`
- Remove no-longer-used `@mixes` / `@polymerMixin` JSDoc annotations
- Stop importing and re-exporting the internal overlay element from the
  root-level entrypoint
