--- packages/breadcrumbs/src/styles/vaadin-breadcrumbs-overlay-base-styles.d.ts
@@ -1,8 +1,0 @@
-/**
- * @license
- * Copyright (c) 2026 - 2026 Vaadin Ltd.
- * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
- */
-import type { CSSResult } from 'lit';
-
-export const breadcrumbsOverlayStyles: CSSResult;
