--- packages/breadcrumbs/src/vaadin-breadcrumbs-overlay-mixin.d.ts
@@ -1,13 +1,0 @@
-/**
- * @license
- * Copyright (c) 2026 - 2026 Vaadin Ltd.
- * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
- */
-import type { Constructor } from '@open-wc/dedupe-mixin';
-import type { PositionMixinClass } from '@vaadin/overlay/src/vaadin-overlay-position-mixin.js';
-
-export declare function BreadcrumbsOverlayMixin<T extends Constructor<HTMLElement>>(
-  base: T,
-): Constructor<BreadcrumbsOverlayMixinClass> & Constructor<PositionMixinClass> & T;
-
-export declare class BreadcrumbsOverlayMixinClass {}
