+++ packages/breadcrumbs/src/vaadin-breadcrumbs-overlay.d.ts
@@ -5,12 +5,12 @@
  */
 import { DirMixin } from '@vaadin/component-base/src/dir-mixin.js';
 import { OverlayMixinClass } from '@vaadin/overlay/src/vaadin-overlay-mixin.js';
-import { BreadcrumbsOverlayMixinClass } from './vaadin-breadcrumbs-overlay-mixin.js';
+import { PositionMixinClass } from '@vaadin/overlay/src/vaadin-overlay-position-mixin.js';
 
 /**
  * An element used internally by `<vaadin-breadcrumbs>`. Not intended to be used separately.
  */
-declare class BreadcrumbsOverlay extends BreadcrumbsOverlayMixinClass(OverlayMixinClass(DirMixin(HTMLElement))) {}
+declare class BreadcrumbsOverlay extends PositionMixinClass(OverlayMixinClass(DirMixin(HTMLElement))) {}
 
 declare global {
   interface HTMLElementTagNameMap {
