+++ packages/breadcrumbs/vaadin-breadcrumbs.js
@@ -1,5 +1,3 @@
-import './src/vaadin-breadcrumbs-overlay.js';
 import './src/vaadin-breadcrumbs.js';
 
 export * from './src/vaadin-breadcrumbs.js';
-export * from './src/vaadin-breadcrumbs-overlay.js';
